remove unused imports from Device::SerialPort
This module doesn't use any of the constants or functions from
Device::SerialPort, so there is no need to import them.

Additionally, the tests use a mock version of the module, which doesn't
provide any of those imports. If any were used, there would be errors
from the test not providing them.

When using the mock module, the unhandled arguments were just ignored by
perl. Future versions of perl are intending to throw errors when
arguments are passed to an undefined import methd.